修复ibu文件

+++ marketing-cs/src/main/java/com/br/marketing/mapper/PhoneSaleIbuMapper.java
@@ -6,5 +6,5 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface PhoneSaleIbuMapper extends PhoneSaleIbuMapperBase {
-    List<PhoneSaleIbu> getPushDassTransferData(@Param("localId") Long localId, @Param("dataId")  Integer dataId);
+    List<PhoneSaleIbu> getPushDassTransferData(@Param("localId") Long localId, @Param("dataId")  Long dataId);
 }
